refactor: Break out network test helper to separate `testing` module
